Refactor separated components
