+++ test/no.2/my-app/src/Components/Board/index.css
@@ -1,0 +1,5 @@
+.board-row:after {
+  clear: both;
+  content: "";
+  display: table;
+}
