+++ test/no.2/my-app/src/Components/Board/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Board, BoardProps } from '../index';
+import Board, { BoardProps } from './index';
 import { Story, Meta } from '@storybook/react/types-6-0';
 import { action } from '@storybook/addon-actions';
