+++ test/no.2/my-app/src/Components/Board/index.tsx
@@ -1,0 +1,37 @@
+import Square from '../Square';
+import './index.css';
+
+type ISquare = '半' | '丁' | null;
+
+export interface BoardProps {
+  squares: ISquare[];
+  onClick: (i: number) => void;
+}
+
+const Board: React.FC<BoardProps> = ({ squares, onClick }) => {
+  const renderSquare = (i: number) => (
+    <Square value={squares[i]} onClick={() => onClick(i)} />
+  );
+
+  return (
+    <div>
+      <div className="board-row">
+        {renderSquare(0)}
+        {renderSquare(1)}
+        {renderSquare(2)}
+      </div>
+      <div className="board-row">
+        {renderSquare(3)}
+        {renderSquare(4)}
+        {renderSquare(5)}
+      </div>
+      <div className="board-row">
+        {renderSquare(6)}
+        {renderSquare(7)}
+        {renderSquare(8)}
+      </div>
+    </div>
+  );
+};
+
+export default Board;
