+++ test/no.2/my-app/src/Components/Game/index.css
@@ -1,0 +1,12 @@
+.game {
+  display: flex;
+  flex-direction: row;
+}
+
+.game-info {
+  margin-left: 20px;
+}
+
+.status {
+  margin-bottom: 10px;
+}
