+++ test/no.2/my-app/src/Components/Game/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Game, GameProps } from '../index';
+import Game, { GameProps } from './index';
 import { Story, Meta } from '@storybook/react/types-6-0';
 
 export default {
