+++ test/no.2/my-app/src/storyshots.test.ts
@@ -1,2 +1,2 @@
 import initStoryshots from '@storybook/addon-storyshots';
-initStoryshots({});
+initStoryshots();
